Move contract and signature stuff to backend
