+++ creusot/src/backend.rs
@@ -3,5 +3,6 @@ pub(crate) mod dependency;
 pub(crate) mod interface;
 pub(crate) mod logic;
 pub(crate) mod program;
+pub(crate) mod signature;
 pub(crate) mod term;
 pub(crate) mod traits;
