+++ creusot/src/util.rs
@@ -1,9 +1,9 @@
 use crate::{
+    backend::signature::sig_to_why3,
     ctx::*,
     translation::{
         pearlite::{super_visit_mut_term, Literal, Term, TermKind, TermVisitorMut},
         specification::PreContract,
-        {self},
     },
 };
 use indexmap::IndexMap;
@@ -28,7 +28,6 @@ use std::{
 use why3::{
     declaration,
     declaration::{LetKind, Signature, ValDecl},
-    exp::Binder,
     Ident, QName,
 };
 
@@ -401,50 +400,6 @@ pub(crate) fn pre_sig_of<'tcx>(
     };
 
     PreSignature { inputs, output, contract }
-}
-
-pub(crate) fn sig_to_why3<'tcx>(
-    ctx: &mut TranslationCtx<'tcx>,
-    names: &mut CloneMap<'tcx>,
-    pre_sig: PreSignature<'tcx>,
-    // FIXME: Get rid of this def id
-    // The PreSig should have the name and the id should be replaced by a param env (if by anything at all...)
-    def_id: DefId,
-) -> Signature {
-    let contract = names.with_public_clones(|names| pre_sig.contract.to_exp(ctx, names));
-
-    let name = item_name(ctx.tcx, def_id, Namespace::ValueNS);
-
-    let span = ctx.tcx.def_span(def_id);
-    let args: Vec<Binder> = names.with_public_clones(|names| {
-        pre_sig
-            .inputs
-            .into_iter()
-            .enumerate()
-            .map(|(ix, (id, _, ty))| {
-                let ty = translation::ty::translate_ty(ctx, names, span, ty);
-                let id = if id.is_empty() {
-                    format!("{}", AnonymousParamName(ix)).into()
-                } else {
-                    ident_of(id)
-                };
-                Binder::typed(id, ty)
-            })
-            .collect()
-    });
-
-    let mut attrs = why3_attrs(ctx.tcx, def_id);
-
-    def_id
-        .as_local()
-        .map(|d| ctx.def_span(d))
-        .and_then(|span| ctx.span_attr(span))
-        .map(|attr| attrs.push(attr));
-
-    let retty = names.with_public_clones(|names| {
-        translation::ty::translate_ty(ctx, names, span, pre_sig.output)
-    });
-    Signature { name, attrs, retty: Some(retty), args, contract }
 }
 
 pub(crate) fn signature_of<'tcx>(
